Reorganize to hide internals, and add start.sh
