+++ src/research_manager.py
