+++ src/strategic_analysis_parser.py
