+++ start.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+if [ ! -d "venv" ]; then
+    python3 -m venv venv
+fi
+
+source venv/bin/activate
+
+pip install -r "src/requirements.txt"
+
+python -m src
+
+deactivate
